requirements.txt
fastapi~=0.115.4
pydantic~=2.9.2
uvicorn~=0.32.0
sqlalchemy~=2.0.36
alembic~=1.14.0
python-dotenv~=1.0.1
psycopg2-binary~=2.9.10
starlette~=0.41.2
pydantic
sqlalchemy
passlib[bcrypt]
